Problem_Statement
The Social Dilemma, a documentary-drama hybrid explores the dangerous human impact of social
networking, with tech experts sounding the alarm on their own creations as the tech experts sound 
the alarm on the dangerous human impact of social networking. This dataset brings you thetwitter 
responses made with the #TheSocialDilemma hashtag after watching the eye-opening
documentary "The Social Dilemma" released in an OTT platform(Netflix) on September 9th, 2020.
You can categorize the tweets into different groups to identify the sentiment of the users regarding
the documentary.
Goal: To categorize the tweets into
three categories.